+++ tests/core/parser/test_get_ignored_files__with_ignored_files.py
@@ -2,7 +2,7 @@
 from assertpy import assert_that
 
 # First party imports
-from printlinter import IgnoreFile, get_ignore_files
+from printlinter import IgnoredFile, get_ignored_files
 
 
 def test_get_ignored_files_with_ignored_files__ALL(
@@ -10,13 +10,13 @@ def test_get_ignored_files_with_ignored_files__ALL(
     testing_files,
 ):
     assert_that(
-        get_ignore_files(
+        get_ignored_files(
             file_with_ignored_all,
             testing_files / "ignored_files/ignore_all.py",
         )
     ).contains_only(
         *[
-            IgnoreFile(
+            IgnoredFile(
                 error_code="ALL",
                 from_file=testing_files / "ignored_files/ignore_all.py",
             )
@@ -29,13 +29,13 @@ def test_get_ignored_files_with_ignored_files__standard_lib(
     testing_files,
 ):
     assert_that(
-        get_ignore_files(
+        get_ignored_files(
             file_with_ignored_standard_lib,
             testing_files / "ignored_files/ignore_ppl000.py",
         )
     ).contains_only(
         *[
-            IgnoreFile(
+            IgnoredFile(
                 error_code="PPL000",
                 from_file=testing_files / "ignored_files/ignore_ppl000.py",
             )
@@ -48,13 +48,13 @@ def test_get_ignored_files_with_ignored_files__print(
     testing_files,
 ):
     assert_that(
-        get_ignore_files(
+        get_ignored_files(
             file_with_ignored_print__file,
             testing_files / "ignored_files/ignore_ppl001.py",
         )
     ).contains_only(
         *[
-            IgnoreFile(
+            IgnoredFile(
                 error_code="PPL001",
                 from_file=testing_files / "ignored_files/ignore_ppl001.py",
             )
@@ -67,13 +67,13 @@ def test_get_ignored_files_with_ignored_files__prettyprint(
     testing_files,
 ):
     assert_that(
-        get_ignore_files(
+        get_ignored_files(
             file_with_ignored_prettyprint__file,
             testing_files / "ignored_files/ignore_ppl002.py",
         )
     ).contains_only(
         *[
-            IgnoreFile(
+            IgnoredFile(
                 error_code="PPL002",
                 from_file=testing_files / "ignored_files/ignore_ppl002.py",
             )
@@ -86,13 +86,13 @@ def test_get_ignored_files_with_ignored_files__stdout_write(
     testing_files,
 ):
     assert_that(
-        get_ignore_files(
+        get_ignored_files(
             file_with_ignored_stdout_write__file,
             testing_files / "ignored_files/ignore_ppl003.py",
         )
     ).contains_only(
         *[
-            IgnoreFile(
+            IgnoredFile(
                 error_code="PPL003",
                 from_file=testing_files / "ignored_files/ignore_ppl003.py",
             )
@@ -105,13 +105,13 @@ def test_get_ignored_files_with_ignored_files__stderr_write(
     testing_files,
 ):
     assert_that(
-        get_ignore_files(
+        get_ignored_files(
             file_with_ignored_stderr_write__file,
             testing_files / "ignored_files/ignore_ppl004.py",
         )
     ).contains_only(
         *[
-            IgnoreFile(
+            IgnoredFile(
                 error_code="PPL004",
                 from_file=testing_files / "ignored_files/ignore_ppl004.py",
             )
@@ -124,13 +124,13 @@ def test_get_ignored_files_with_ignored_files__stdout_writelines(
     testing_files,
 ):
     assert_that(
-        get_ignore_files(
+        get_ignored_files(
             file_with_ignored_stdout_writelines__file,
             testing_files / "ignored_files/ignore_ppl005.py",
         )
     ).contains_only(
         *[
-            IgnoreFile(
+            IgnoredFile(
                 error_code="PPL005",
                 from_file=testing_files / "ignored_files/ignore_ppl005.py",
             )
@@ -143,13 +143,13 @@ def test_get_ignored_files_with_ignored_files__stderr_writelines(
     testing_files,
 ):
     assert_that(
-        get_ignore_files(
+        get_ignored_files(
             file_with_ignored_stderr_writelines__file,
             testing_files / "ignored_files/ignore_ppl006.py",
         )
     ).contains_only(
         *[
-            IgnoreFile(
+            IgnoredFile(
                 error_code="PPL006",
                 from_file=testing_files / "ignored_files/ignore_ppl006.py",
             )
@@ -162,7 +162,7 @@ def test_get_ignored_files_with_ignored_files__comment_in_wrong_place(
     testing_files,
 ):
     assert_that(
-        get_ignore_files(
+        get_ignored_files(
             file_with_comment_in_wrong_place__file,
             testing_files / "ignored_files/disable_in_wrong_place.py",
         )
@@ -174,13 +174,13 @@ def test_get_ignored_files_with_ignored_files__comment_top_of_the_file_not_first
     testing_files,
 ):
     assert_that(
-        get_ignore_files(
+        get_ignored_files(
             file_with_comment_not_at_first_line__file,
             testing_files / "ignored_files/disable_valid_but_not_in_first_line.py",
         )
     ).contains_only(
         *[
-            IgnoreFile(
+            IgnoredFile(
                 error_code="ALL",
                 from_file=testing_files
                 / "ignored_files/disable_valid_but_not_in_first_line.py",