+++ tests/core/parser/test_get_ignored_lines__without_ignored_lines.py
@@ -2,7 +2,7 @@
 from assertpy import assert_that
 
 # First party imports
-from printlinter import get_ignore_lines
+from printlinter import get_ignored_lines
 
 
 def test_get_ignored_lines_without_ignored_lines__print(
@@ -10,7 +10,7 @@ def test_get_ignored_lines_without_ignored_lines__print(
     testing_files,
 ):
     assert_that(
-        get_ignore_lines(file_without_ignored_print, testing_files / "print/toto_1.py")
+        get_ignored_lines(file_without_ignored_print, testing_files / "print/toto_1.py")
     ).is_equal_to([])
 
 
@@ -19,7 +19,7 @@ def test_get_ignored_lines_without_ignored_lines__prettyprint(
     testing_files,
 ):
     assert_that(
-        get_ignore_lines(
+        get_ignored_lines(
             file_without_ignored_prettyprint,
             testing_files / "pprint/pprint1.py",
         )
@@ -31,7 +31,7 @@ def test_get_ignored_lines_without_ignored_lines__sys_stdout_write(
     testing_files,
 ):
     assert_that(
-        get_ignore_lines(
+        get_ignored_lines(
             file_without_ignored_stdout_write,
             testing_files / "sys/stdout/write/stdout1.py",
         )
@@ -43,7 +43,7 @@ def test_get_ignored_lines_without_ignored_lines__sys_stderr_write(
     testing_files,
 ):
     assert_that(
-        get_ignore_lines(
+        get_ignored_lines(
             file_without_ignored_stderr_write,
             testing_files / "sys/stderr/write/stderr1.py",
         )
@@ -55,7 +55,7 @@ def test_get_ignored_lines_without_ignored_lines__sys_stdout_writelines(
     testing_files,
 ):
     assert_that(
-        get_ignore_lines(
+        get_ignored_lines(
             file_without_ignored_stdout_writelines,
             testing_files / "sys/stdout/writelines/stdout1.py",
         )
@@ -67,7 +67,7 @@ def test_get_ignored_lines_without_ignored_lines__sys_stderr_writelines(
     testing_files,
 ):
     assert_that(
-        get_ignore_lines(
+        get_ignored_lines(
             file_without_ignored_stderr_writelines,
             testing_files / "sys/stderr/writelines/stderr1.py",
         )