+++ tests/core/test_classes.py
@@ -6,7 +6,7 @@ from pytest import param
 from assertpy import assert_that, soft_assertions
 
 # First party imports
-from printlinter import IgnoredBlock, IgnoreFile, IgnoreLine, IssueEnum, IssueInfo
+from printlinter import IgnoredBlock, IgnoredFile, IgnoredLine, IssueEnum, IssueInfo
 
 
 @pytest.mark.parametrize(
@@ -100,7 +100,7 @@ def test_issue_info_class(line, column, str_line, file, ignored, issue_type):
         num_col=column,
         line_as_str=str_line,
         from_file=file,
-        ignore=ignored,
+        ignored=ignored,
     )
     with soft_assertions():
         assert issue
@@ -129,8 +129,8 @@ def test_issue_info_class(line, column, str_line, file, ignored, issue_type):
         param("PPL006", id="sys.stderr.writelines PPL006"),
     ],
 )
-def test_ignore_line_class(file, err_code):
-    assert IgnoreLine(1, err_code, file)
+def test_ignored_line_class(file, err_code):
+    assert IgnoredLine(1, err_code, file)
 
 
 @pytest.mark.parametrize(
@@ -151,8 +151,8 @@ def test_ignore_line_class(file, err_code):
         param("PPL006", id="sys.stderr.writelines PPL006"),
     ],
 )
-def test_ignore_file_class(file, err_code):
-    assert IgnoreFile(err_code, file)
+def test_ignored_file_class(file, err_code):
+    assert IgnoredFile(err_code, file)
 
 
 @pytest.mark.parametrize(