+++ tests/testing_files/ignored_block/no_re_enable/ppl002.py
@@ -6,6 +6,6 @@ def toto():
     return 1 + 2
 
 
-# <py-printlinter disable PPL002>
+# <printlinter disable PPL002>
 pprint(toto())
 pprint(toto() + toto())