+++ tests/testing_files/ignored_block/no_re_enable/ppl005.py
@@ -6,6 +6,6 @@ def toto():
     return 1 + 2
 
 
-# <py-printlinter disable PPL005>
+# <printlinter disable PPL005>
 stdout.writelines(toto())
 stdout.writelines(toto() + toto())