+++ tests/testing_files/ignored_block/no_re_enable/ppl006.py
@@ -6,6 +6,6 @@ def toto():
     return 1 + 2
 
 
-# <py-printlinter disable PPL006>
+# <printlinter disable PPL006>
 stderr.writelines(toto())
 stderr.writelines(toto() + toto())