+++ tests/testing_files/ignored_block/re_enable/all.py
@@ -7,11 +7,11 @@ def toto():
     return 1 + 2
 
 
-# <py-printlinter disable ALL>
+# <printlinter disable ALL>
 print(toto())
 pprint(toto())
 stdout.write(toto())
 stderr.write(toto())
 stdout.writelines(toto())
 stderr.writelines(toto())
-# <py-printlinter enable ALL>
+# <printlinter enable ALL>