+++ tests/testing_files/ignored_block/re_enable/ppl001.py
@@ -2,7 +2,7 @@ def toto():
     return 1 + 2
 
 
-# <py-printlinter disable PPL001>
+# <printlinter disable PPL001>
 print(toto())
 print(toto() + toto())
-# <py-printlinter enable PPL001>
+# <printlinter enable PPL001>